services/upload/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
pydantic-settings==2.10.1
